Implement a Load Benchmark Runner with no flow control on scan
Due to large perfomance losses, choosing a ready channel
to send scanned batches to, when using ~100 channels
a separate implementation of the benchmark runner uses
a simplified algorithm to distirbute the scanned batches.
The algorithm follows how Prometheus distributes batches to
write to remote storage. Benchmarks show that it works better
for all kind of batch sizes. And drasticly improves performance
for large number of channels.

Also:
  - Wrote test for the scan with no flow controll
  - Added flow-control flag to tsbs_load

+++ cmd/tsbs_load_cratedb/scan_test.go
@@ -63,7 +63,7 @@ func TestEventsBatch(t *testing.T) {
 	for _, p := range points {
 		eb.Append(p)
 	}
-	if eb.Len() != len(points) {
+	if eb.Len() != uint(len(points)) {
 		t.Errorf(fmt.Sprintf("eventsBatch must have %d points", len(points)))
 	}
 	if len(eb.batches) != 2 {
